refactor : consts 패키지 생성 (#559)
