+++ src/main/java/com/clubber/ClubberServer/domain/auth/controller/AuthController.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.auth.controller;
 
-import static com.clubber.ClubberServer.global.common.ClubberStatic.*;
+import static com.clubber.ClubberServer.global.common.consts.ClubberStatic.*;
 
 import com.clubber.ClubberServer.domain.auth.component.UserRegister;
 import com.clubber.ClubberServer.domain.auth.component.UserWithdraw;
